부딪힘(기반)
