+++ Assets/POLYGONCityCharacters/Prefabs/Pedestrians/Biker_pede.prefab
@@ -1,6 +1,6 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!1 &1912664198549763251
+--- !u!1 &1382111985350901868
 GameObject:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -8,8 +8,8 @@ GameObject:
   m_PrefabAsset: {fileID: 0}
   serializedVersion: 6
   m_Component:
-  - component: {fileID: 3212421168268600982}
-  - component: {fileID: 648323271784283348}
+  - component: {fileID: 1760676423933131511}
+  - component: {fileID: 7944542060528789165}
   m_Layer: 6
   m_Name: FrontTrigger
   m_TagString: Untagged
@@ -17,13 +17,13 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &3212421168268600982
+--- !u!4 &1760676423933131511
 Transform:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1912664198549763251}
+  m_GameObject: {fileID: 1382111985350901868}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 1.8967767, z: 0}
@@ -32,13 +32,13 @@ Transform:
   m_Children: []
   m_Father: {fileID: 6796165826500184900}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!65 &648323271784283348
+--- !u!65 &7944542060528789165
 BoxCollider:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1912664198549763251}
+  m_GameObject: {fileID: 1382111985350901868}
   m_Material: {fileID: 0}
   m_IncludeLayers:
     serializedVersion: 2
@@ -64,6 +64,10 @@ PrefabInstance:
     - target: {fileID: 172216, guid: b4587275bd8858843b068385a1cb3677, type: 3}
       propertyPath: m_Name
       value: Biker_pede
+      objectReference: {fileID: 0}
+    - target: {fileID: 172216, guid: b4587275bd8858843b068385a1cb3677, type: 3}
+      propertyPath: m_TagString
+      value: Pedestrian
       objectReference: {fileID: 0}
     - target: {fileID: 485478, guid: b4587275bd8858843b068385a1cb3677, type: 3}
       propertyPath: m_LocalScale.x
@@ -138,7 +142,7 @@ PrefabInstance:
     m_AddedGameObjects:
     - targetCorrespondingSourceObject: {fileID: 485478, guid: b4587275bd8858843b068385a1cb3677, type: 3}
       insertIndex: -1
-      addedObject: {fileID: 3212421168268600982}
+      addedObject: {fileID: 1760676423933131511}
     m_AddedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: b4587275bd8858843b068385a1cb3677, type: 3}
 --- !u!4 &6796165826500184900 stripped
