+++ Assets/POLYGONCityCharacters/Prefabs/Pedestrians/FastFoodGuy_Pede.prefab
@@ -86,7 +86,7 @@ Transform:
   - {fileID: 4224215181262901454}
   m_Father: {fileID: 3631090995250120637}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!1 &5070085816885808235
+--- !u!1 &6618341470275273678
 GameObject:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -94,8 +94,8 @@ GameObject:
   m_PrefabAsset: {fileID: 0}
   serializedVersion: 6
   m_Component:
-  - component: {fileID: 8205524399235378344}
-  - component: {fileID: 5063854742937074552}
+  - component: {fileID: 3494089598814420505}
+  - component: {fileID: 6189680216479314141}
   m_Layer: 6
   m_Name: FrontTrigger
   m_TagString: Untagged
@@ -103,13 +103,13 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &8205524399235378344
+--- !u!4 &3494089598814420505
 Transform:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5070085816885808235}
+  m_GameObject: {fileID: 6618341470275273678}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 1.8967767, z: 0}
@@ -118,13 +118,13 @@ Transform:
   m_Children: []
   m_Father: {fileID: 3631090995250120637}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!65 &5063854742937074552
+--- !u!65 &6189680216479314141
 BoxCollider:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5070085816885808235}
+  m_GameObject: {fileID: 6618341470275273678}
   m_Material: {fileID: 0}
   m_IncludeLayers:
     serializedVersion: 2
@@ -261,6 +261,10 @@ PrefabInstance:
     - target: {fileID: 130974, guid: cf0f33d2a32f159499351bdc8d15d958, type: 3}
       propertyPath: m_Layer
       value: 6
+      objectReference: {fileID: 0}
+    - target: {fileID: 130974, guid: cf0f33d2a32f159499351bdc8d15d958, type: 3}
+      propertyPath: m_TagString
+      value: Pedestrian
       objectReference: {fileID: 0}
     - target: {fileID: 131236, guid: cf0f33d2a32f159499351bdc8d15d958, type: 3}
       propertyPath: m_Layer
@@ -474,7 +478,7 @@ PrefabInstance:
       addedObject: {fileID: 9094948884135977140}
     - targetCorrespondingSourceObject: {fileID: 412478, guid: cf0f33d2a32f159499351bdc8d15d958, type: 3}
       insertIndex: -1
-      addedObject: {fileID: 8205524399235378344}
+      addedObject: {fileID: 3494089598814420505}
     m_AddedComponents:
     - targetCorrespondingSourceObject: {fileID: 130974, guid: cf0f33d2a32f159499351bdc8d15d958, type: 3}
       insertIndex: -1
