+++ Assets/POLYGONCityCharacters/Prefabs/Pedestrians/HipsterGirl_pede.prefab
@@ -141,7 +141,7 @@ CapsuleCollider:
   m_Height: 1.87
   m_Direction: 1
   m_Center: {x: 0, y: 0.96, z: 0}
---- !u!1 &6664889449078749137
+--- !u!1 &8176480486688710977
 GameObject:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -149,8 +149,8 @@ GameObject:
   m_PrefabAsset: {fileID: 0}
   serializedVersion: 6
   m_Component:
-  - component: {fileID: 6096906675187104335}
-  - component: {fileID: 6116055272234699375}
+  - component: {fileID: 889444650767192212}
+  - component: {fileID: 4330411085297758043}
   m_Layer: 6
   m_Name: FrontTrigger
   m_TagString: Untagged
@@ -158,13 +158,13 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &6096906675187104335
+--- !u!4 &889444650767192212
 Transform:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6664889449078749137}
+  m_GameObject: {fileID: 8176480486688710977}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 1.8967767, z: 0}
@@ -173,13 +173,13 @@ Transform:
   m_Children: []
   m_Father: {fileID: 6800029457925334794}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!65 &6116055272234699375
+--- !u!65 &4330411085297758043
 BoxCollider:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6664889449078749137}
+  m_GameObject: {fileID: 8176480486688710977}
   m_Material: {fileID: 0}
   m_IncludeLayers:
     serializedVersion: 2
@@ -382,6 +382,10 @@ PrefabInstance:
       propertyPath: m_Layer
       value: 6
       objectReference: {fileID: 0}
+    - target: {fileID: 178284, guid: fe631d66d91635d4db4b17b1c3da562a, type: 3}
+      propertyPath: m_TagString
+      value: Pedestrian
+      objectReference: {fileID: 0}
     - target: {fileID: 180794, guid: fe631d66d91635d4db4b17b1c3da562a, type: 3}
       propertyPath: m_Layer
       value: 6
@@ -492,7 +496,7 @@ PrefabInstance:
       addedObject: {fileID: 2762882740218784777}
     - targetCorrespondingSourceObject: {fileID: 486310, guid: fe631d66d91635d4db4b17b1c3da562a, type: 3}
       insertIndex: -1
-      addedObject: {fileID: 6096906675187104335}
+      addedObject: {fileID: 889444650767192212}
     m_AddedComponents:
     - targetCorrespondingSourceObject: {fileID: 178284, guid: fe631d66d91635d4db4b17b1c3da562a, type: 3}
       insertIndex: -1
