--- Assets/Scripts/IntersectionPoolManager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 8e0456322ee9e1d4597bc22b5464e0bf
