--- Assets/Scripts/PedestrainSpawner.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 55170639d06bc7544b9a813a49afe79e
