--- Assets/Scripts/PedestrianAnimatorController.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 5e8b2f7227c58de418757c1d63f2328d
