--- Assets/Scripts/PedestrianMovement.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 9ac310136a7dafe488b251aaf16463e6
