--- Assets/Scripts/PedestrianPool.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 97e84fd2a79f4874fbada55465c02c37
