+++ Assets/Scripts/Sh.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 081d7cff636e6194eb97fa8b70217c9b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
