--- Assets/Scripts/WallGenerator.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: c34ba20f5dd71504f98fb3a3f848dda1
