+++ ProjectSettings/TagManager.asset
@@ -14,6 +14,8 @@ TagManager:
   - Terrain
   - Node
   - Cell
+  - Pedestrian
+  - Building
   layers:
   - Default
   - TransparentFX
